requirements.txt
aniso8601==9.0.1
atlassian-python-api==3.17.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
Deprecated==1.2.13
Flask==2.0.2
Flask-REST==1.3
flask-restx==0.5.1
idna==3.3
importlib-metadata==4.10.0
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.3.3
MarkupSafe==2.0.1
oauthlib==3.1.1
pyrsistent==0.18.0
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
slack-sdk==3.13.0
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.7.0
